refactor: shared infer, extract validator guards
