+++ packages/pure-parse/src/validate/index.ts
@@ -1,2 +1,3 @@
 export * from './validation'
 export * from './json'
+export * from './guards'
